+++ backend/api/admin/src/test/java/co/airy/core/api/admin/TemplatesControllerTest.java
@@ -81,7 +81,7 @@ public class TemplatesControllerTest {
         final String content = "{\"blueprint\":\"text\",\"payload\":\"[[salutation]]!\"}";
         final String payload = "{\"name\":\"" + name + "\",\"source\": \"" + source + "\",\"content\":" + content + "}";
 
-        final String createTemplateResponse = webTestHelper.post("/templates.create", payload, "user-id")
+        final String createTemplateResponse = webTestHelper.post("/templates.create", payload)
                 .andExpect(status().isCreated())
                 .andReturn()
                 .getResponse()
@@ -92,25 +92,25 @@ public class TemplatesControllerTest {
 
         assertThat(templateId, is(not(nullValue())));
 
-        retryOnException(() -> webTestHelper.post("/templates.info", "{\"id\":\"" + templateId + "\"}", "user-id")
+        retryOnException(() -> webTestHelper.post("/templates.info", "{\"id\":\"" + templateId + "\"}")
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.id").value(is(templateId)))
                 .andExpect(jsonPath("$.content.blueprint").value(is("text")))
                 .andExpect(jsonPath("$.source").value(is(source)))
                 .andExpect(jsonPath("$.name").value(is(name))), "could not find template");
 
-        webTestHelper.post("/templates.update", "{\"id\":\"" + templateId + "\", \"name\": \"new-template-name\", \"source\": \"google\", \"content\": " + content + "}", "user-id")
+        webTestHelper.post("/templates.update", "{\"id\":\"" + templateId + "\", \"name\": \"new-template-name\", \"source\": \"google\", \"content\": " + content + "}")
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.id").value(is(templateId)))
                 .andExpect(jsonPath("$.content.blueprint").value(is("text")))
                 .andExpect(jsonPath("$.source").value(is("google")))
                 .andExpect(jsonPath("$.name").value(is("new-template-name")));
 
-        retryOnException(() -> webTestHelper.post("/templates.info", "{\"id\":\"" + templateId + "\"}", "user-id")
+        retryOnException(() -> webTestHelper.post("/templates.info", "{\"id\":\"" + templateId + "\"}")
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.name").value(is("new-template-name"))), "could not update template");
 
-        webTestHelper.post("/templates.delete", "{\"id\": \"" + templateId + "\"}", "user-id").andExpect(status().isNoContent());
+        webTestHelper.post("/templates.delete", "{\"id\": \"" + templateId + "\"}").andExpect(status().isNoContent());
     }
 
     @Test
@@ -122,7 +122,7 @@ public class TemplatesControllerTest {
             final String name = "awesome-template-" + i;
             final String payload = "{\"name\":\"" + name + "\",\"source\": \"" + googleSource + "\",\"content\":" + googleContent + "}";
 
-            webTestHelper.post("/templates.create", payload, "user-id").andExpect(status().isCreated());
+            webTestHelper.post("/templates.create", payload).andExpect(status().isCreated());
         }
 
         final String twilioSource = "twilio.sms";
@@ -131,36 +131,36 @@ public class TemplatesControllerTest {
             final String name = "awesome-template-" + i;
             final String payload = "{\"name\":\"" + name + "\",\"source\": \"" + twilioSource + "\",\"content\":" + twilioContent + "}";
 
-            webTestHelper.post("/templates.create", payload, "user-id").andExpect(status().isCreated());
+            webTestHelper.post("/templates.create", payload).andExpect(status().isCreated());
         }
 
         retryOnException(() -> webTestHelper.post("/templates.list",
-                "{\"source\": \"google\"}", "user-id")
+                "{\"source\": \"google\"}")
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.data.length()", is(10))), "could not list google templates");
 
         retryOnException(() -> webTestHelper.post("/templates.list",
-                "{\"source\": \"google\", \"name\": \"awesome-template-\"}", "user-id")
+                "{\"source\": \"google\", \"name\": \"awesome-template-\"}")
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.data.length()", is(10))), "could not list google templates by name");
 
         retryOnException(() -> webTestHelper.post("/templates.list",
-                "{\"source\": \"google\", \"name\": \"awesome-template-1\"}", "user-id")
+                "{\"source\": \"google\", \"name\": \"awesome-template-1\"}")
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.data.length()", is(1))), "could not list google templates by name");
 
         retryOnException(() -> webTestHelper.post("/templates.list",
-                "{\"source\": \"twilio.sms\"}", "user-id")
+                "{\"source\": \"twilio.sms\"}")
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.data.length()", is(5))), "could not list twilio templates");
 
         retryOnException(() -> webTestHelper.post("/templates.list",
-                "{\"source\": \"twilio.sms\", \"name\": \"awesome-template-\"}", "user-id")
+                "{\"source\": \"twilio.sms\", \"name\": \"awesome-template-\"}")
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.data.length()", is(5))), "could not list twilio templates by name");
 
         retryOnException(() -> webTestHelper.post("/templates.list",
-                "{\"source\": \"twilio.sms\", \"name\": \"awesome-template-1\"}", "user-id")
+                "{\"source\": \"twilio.sms\", \"name\": \"awesome-template-1\"}")
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.data.length()", is(1))), "could not list twilio templates by name");
     }