+++ frontend/ui/src/InitializeAiryApi.ts
@@ -1,14 +1,6 @@
 import {HttpClient} from 'httpclient';
-import {getAuthToken} from './cookies';
 import {env} from './env';
-import {store} from './store';
-import {logoutUser} from './actions/user';
 
-const authToken = getAuthToken();
-
-export const HttpClientInstance = new HttpClient(env.API_HOST, authToken, error => {
-  console.error('Unauthorized request, logging out user');
+export const HttpClientInstance = new HttpClient(env.API_HOST, error => {
   console.error(error);
-
-  logoutUser()(store.dispatch);
 });