+++ integration/chat-plugin/websocket_test.spec.ts
@@ -9,16 +9,6 @@ import {
 
 describe('Websocket test', () => {
   it('Send message from Inbox to Chatplugin and assert Websocket is working', () => {
-    cy.visit('/ui/login');
-    cy.get('form')
-      .within(() => {
-        cy.get('input[type=email]').type(Cypress.env('username'));
-        cy.get('input[type=password]').type(Cypress.env('password'));
-      })
-      .submit();
-
-    cy.wait(500);
-
     cy.visit('/ui/channels');
     cy.wait(500);
     cy.url().should('include', '/channels');
@@ -38,46 +28,37 @@ describe('Websocket test', () => {
 
     cy.wait(2500);
 
-    cy.request('POST', '/users.login', {
-      email: Cypress.env('username'),
-      password: Cypress.env('password'),
+    cy.request({
+      method: 'POST',
+      url: '/conversations.list',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: {
+        cursor: null,
+        page_size: 2,
+      },
     }).then(response => {
-      const loginToken = response.body['token'];
+      const conversationId = response.body.data[0].id;
 
       cy.request({
         method: 'POST',
-        url: '/conversations.list',
+        url: '/messages.send',
         headers: {
           'Content-Type': 'application/json',
-          Authorization: `Bearer ${loginToken}`,
         },
         body: {
-          cursor: null,
-          page_size: 2,
+          conversation_id: `${conversationId}`,
+          message: {
+            text: Cypress.env('websocketMessage'),
+          },
         },
       }).then(response => {
-        const conversationId = response.body.data[0].id;
-
-        cy.request({
-          method: 'POST',
-          url: '/messages.send',
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${loginToken}`,
-          },
-          body: {
-            conversation_id: `${conversationId}`,
-            message: {
-              text: Cypress.env('websocketMessage'),
-            },
-          },
-        }).then(response => {
-          expect(response).property('status').to.equal(200);
-        });
-        cy.wait(3000);
-
-        cy.get(`[data-cy=${cyChatPluginMessageList}]`).children().its('length').should('eq', 3);
+        expect(response).property('status').to.equal(200);
       });
+      cy.wait(3000);
+
+      cy.get(`[data-cy=${cyChatPluginMessageList}]`).children().its('length').should('eq', 3);
     });
   });
 });